Progress bar

+++ db/migrate/20200308180511_add_okr_id_to_milestones.rb
@@ -1,5 +1,5 @@
 class AddOkrIdToMilestones < ActiveRecord::Migration[6.0]
   def change
-    add_column :milestones, :okr_id, :integer
+    add_column :milestones, :okr_id, :bigint
   end
 end
